move: move to util/lock
