+++ app/src/main/java/com/nicer/attiary/util/lock/AppLock.kt
@@ -1,4 +1,5 @@
-package com.nicer.attiary.data.app
+package com.nicer.attiary.util.lock
+
 import android.content.Context
 
 class AppLock(context: Context) {
@@ -6,16 +7,16 @@ class AppLock(context: Context) {
 	private var sharedPref = context.getSharedPreferences("appLock", Context.MODE_PRIVATE)
 
 	// 잠금 설정
-	fun setPassLock(password : String){
-		sharedPref.edit().apply{
+	fun setPassLock(password: String) {
+		sharedPref.edit().apply {
 			putString("applock", password)
 			apply()
 		}
 	}
 
 	// 잠금 설정 제거
-	fun removePassLock(){
-		sharedPref.edit().apply{
+	fun removePassLock() {
+		sharedPref.edit().apply {
 			remove("applock")
 			apply()
 		}
@@ -23,12 +24,12 @@ class AppLock(context: Context) {
 
 	// 입력한 비밀번호가 맞는가?
 	fun checkPassLock(password: String): Boolean {
-		return sharedPref.getString("applock","0") == password
+		return sharedPref.getString("applock", "0") == password
 	}
 
 	// 잠금 설정이 되어있는가?
 	fun isPassLockSet(): Boolean {
-		if(sharedPref.contains("applock")){
+		if (sharedPref.contains("applock")) {
 			return true
 		}
 		return false
