+++ app/src/main/java/com/nicer/attiary/util/lock/AppLockStatus.kt
@@ -1,9 +1,10 @@
-package com.nicer.attiary.data.app
+package com.nicer.attiary.util.lock
 
 var lock = true
-class AppLockConst {
-	object AppLockCosnt{
-		val type ="type"
+
+class AppLockStatus {
+	object AppLockStatus {
+		val TYPE = "type"
 		val ENABLE_PASSLOCK = 1 //잠금설정
 		val DISABLE_PASSLOCK = 2 //잠금 비활성화
 		val CHANGE_PASSWORD = 3 //암호변경
